Fixed some bugs. Tested for all Working examples - our assembler generates the same output as MARS. So, for any valid code our program assembles without error and generates the correct output. Now we need to verify that for all invalid codes our program signs the correct error.
Also build the comparer program, who compares txt windows and unix files and say if they have differences

+++ src/MIPS Assembler/code_generator.c
@@ -38,7 +38,7 @@ int code_generator(inst_list* i_list, FILE* output_file) {
 				inst += (p->values).op 		<< 26;
 				inst += (p->values).rs 		<< 21;
 				inst += (p->values).rt 		<< 16;
-				inst += (p->values).imm 	<< 0;
+				inst += (uint16_t)(p->values).imm 	<< 0;
 				break;
 			case TYPE_J:
 				inst += (p->values).op 		<< 26;
